refactor(CompliancePolicies): divide into two files
