+++ src/SmartComponents/CompliancePolicies/CompliancePoliciesGraphQL.js
@@ -1,0 +1,90 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import { useQuery } from '@apollo/client';
+import { Grid } from '@patternfly/react-core';
+import PageHeader, {
+  PageHeaderTitle,
+} from '@redhat-cloud-services/frontend-components/PageHeader';
+import ComplianceEmptyState from 'PresentationalComponents/ComplianceEmptyState';
+import {
+  LinkWithPermission as Link,
+  ErrorPage,
+  LoadingPoliciesTable,
+  PoliciesTable,
+  StateView,
+  StateViewPart,
+  LinkButton,
+} from 'PresentationalComponents';
+
+import { QUERY } from './constants';
+
+const CompliancePoliciesGraphQL = () => {
+  let { data, error, loading } = useQuery(QUERY);
+
+  const CreateLink = () => (
+    <Link
+      to="/scappolicies/new"
+      Component={LinkButton}
+      componentProps={{
+        variant: 'primary',
+        ouiaId: 'CreateNewPolicyButton',
+      }}
+    >
+      Create new policy
+    </Link>
+  );
+
+  let policies;
+
+  if (data) {
+    error = undefined;
+    loading = undefined;
+    if (data) {
+      policies = data.profiles.edges.map((profile) => profile.node);
+    } else {
+      data = [];
+      policies = [];
+    }
+  }
+
+  return (
+    <React.Fragment>
+      <PageHeader className="page-header">
+        <PageHeaderTitle title="SCAP policies" />
+      </PageHeader>
+      <section className="pf-v5-c-page__main-section">
+        <StateView stateValues={{ error, data, loading }}>
+          <StateViewPart stateKey="error">
+            <ErrorPage error={error} />
+          </StateViewPart>
+          <StateViewPart stateKey="loading">
+            <LoadingPoliciesTable />
+          </StateViewPart>
+          <StateViewPart stateKey="data">
+            {policies && policies.length === 0 ? (
+              <Grid hasGutter>
+                <ComplianceEmptyState
+                  title="No policies"
+                  mainButton={<CreateLink />}
+                />
+              </Grid>
+            ) : (
+              <PoliciesTable policies={policies} DedicatedAction={CreateLink} />
+            )}
+          </StateViewPart>
+        </StateView>
+      </section>
+    </React.Fragment>
+  );
+};
+
+CompliancePoliciesGraphQL.propTypes = {
+  query: PropTypes.shape({
+    data: PropTypes.object,
+    error: PropTypes.string,
+    loading: PropTypes.bool,
+    refetch: PropTypes.func,
+  }),
+};
+
+export default CompliancePoliciesGraphQL;
